+++ src/molecules/content/Label.tsx
@@ -7,6 +7,7 @@ type LabelProps = {
   text: string;
   style?: ViewStyle | RegisteredStyle<ViewStyle>;
   textColor?: string;
+  numberOfLines?: 1 | 2;
 };
 
 const styles = StyleSheet.create({
@@ -22,7 +23,10 @@ const textStyle = registeredTextStyle('title', colors.uma, 'Label');
 
 export const LabelUnmemoized = (props: LabelProps) => (
   <View style={[styles.Wrapper, props.style]}>
-    <Text style={[textStyle, { color: props.textColor || colors.uma }]} numberOfLines={1}>
+    <Text
+      style={[textStyle, { color: props.textColor || colors.uma }]}
+      numberOfLines={props.numberOfLines ?? 1}
+    >
       {props.text}
     </Text>
   </View>